ProductRestController.deleteById returns http status OK in case of success and NOT FOUND in case of failure

+++ src/main/java/com/smalaca/restapp/api/rest/product/ProductRestController.java
@@ -59,8 +59,13 @@ public class ProductRestController {
     }
 
     @DeleteMapping("{id}")
-    public void deleteById(@PathVariable Long id) {
-        repository.deleteById(id);
+    public ResponseEntity<Void> deleteById(@PathVariable Long id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return ResponseEntity.ok().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     @PostMapping
